added module template for pallets needed
